--- src/preset/node/utils/index.ts
@@ -1,2 +1,0 @@
-export { axiosTracingMiddleware } from './http-client';
-export { getClientIp } from './http-server';
